+++ src/Gate.IO.Api/Futures/Responses/GateFuturesChaseOrderCreateResponse.cs
@@ -1,0 +1,7 @@
+namespace Gate.IO.Api.Futures;
+
+internal record GateFuturesChaseOrderCreateResponse
+{
+    [JsonProperty("id")]
+    public string OrderId { get; set; }
+}