+++ src/Gate.IO.Api/Futures/Responses/GateFuturesChaseOrderDetailResponse.cs
@@ -1,0 +1,7 @@
+namespace Gate.IO.Api.Futures;
+
+internal record GateFuturesChaseOrderDetailResponse
+{
+    [JsonProperty("order")]
+    public GateFuturesChaseOrder Order { get; set; }
+}